refactor: delete_event.ts を delete_recruit/ に移動
